requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0
torchmetrics>=0.11.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
hydra-optuna-sweeper>=1.2.0
pyyaml>=6.0
tqdm>=4.64.0
hydra-joblib-launcher
plotly>=5.0.0
scipy>=1.7.0